requirements.txt
fastapi
SQLAlchemy
psycopg2-binary
python-jose
ipykernel
firebase-admin
uvicorn
python-multipart
firebase_admin
python-dotenv
pydantic_settings
google-adk
litellm
google-generativeai